+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Our Keycloak version is working well with PostgreSQL database. For using other S
 - OID-FED settings for OPs
 - Update Identity Provider in explicit registration (OpenId Federation as RP)
 - Support explicit registration with Trust Chains as OP
+- Refresh for OpenID Federation Identity Provider
 
 ### Changed
 - Move entityType and ClientRegistrationType to OpenId Federation general settings